Improved login and added logout
I made login to go to admin.java. In there it checks if there is a session and if not it throws you back to login page. Also if trying to login and no password has been set it redirects back to login page. Importantly also added logging out, still in progress to be improved on.