Added deleting and updating/editing questions
I added the deletion of the questions. Also I added the updating/editing of the questions. Both are handled by the id of the question and are heavily checked that there is and id so nothing should break. This means you can't just type in the url you need to press to button (you can do it in just url but this just makes sure that it doesn't break anything).